Add headroom.js to clean code in the main.js
In this commit we added headroom.js to get rid of all the nested ifs in
the main.js file. Headroom has the functionality to bind to events when
the user scrolls down or up the page, also has events for when the page
is at its top or at its bottom.

We modified the styles from the mobile-menu.scss so the classes override
correctly when were on top or bottom of the page.