Added hire-us page with contact form
We are submitting form to formspree.io, and we used
the `_next` option to specify a custom thanks page.
Validations are being handled on form submittion, and we
show a little required warning next to each input.
Also made sure that the contact form validation script
only gets included on the hire-us page.