Added some webpack plugins to optimize assets
Optimizations added:
* Added image optimization with `imagemin-webpack-plugin`
* Added PurifyCSS with `purifycss-webpack`
* Added ES6 support, and updated some scripts to use ES6 features

Also, all assets now live on the `_assets/` folder.

With purifyCSS we managed to reduce `main.css` bundle size
from ~180kb to ~120kb.

Aside from this fixed a bug with the `_next` url for the hire-us form,
as it wasn't working on languages other than the default one.